pp_ocr_service/requirements.txt
fastapi>=0.100.0
uvicorn[standard]>=0.22.0
python-multipart>=0.0.6
paddleocr>=2.7.0
paddlepaddle>=2.5.0
Pillow>=9.0.0
numpy>=1.21.0
